requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
pydantic-settings